Add support for returning the length of the input Color Tag
This is the final piece to add support for a single pass over the Color Tag, this new method is naive like the rest of the generator in that the tags should be valid, as such it returns the expected length of the Color Tag + it's terminator.

So for a tag of "{red}" it will return 4, as "red}" will have been passed for matching.